refactor: rename constants files for consistency
